Load gpx from a text area and sort by time.
This now loads GPX from XML and displays all WPTs to the map.  I need to figure out a subset of valid GPX to focus on, to make it easier to display in a reasonable way.  I'll probably start with tracks (to display data such as runner positions over the course of a race).